Exclude attachment with write mask=none from render pass in WebGL compatible mode.
Because WebGL allows shader to write incompatible data type if the attachment
is disabled. In this case, we need to exclude the attachment from metal
render pass since metal doesn't allow incompatible shader write even if
write mask=none.

This has a down side of new render pass has to be started whenever an
attachment is disabled/enabled.

Change-Id: I0a67ca634756e893d1b4c6a61cb84856711ca855